fix(clippy): sort by key, and drop a u64 cast on a u64
`cargo clippy --workspace --all-targets -- -D warnings` is a CI gate, and
two lints landed with this session's work: a comparator that could be a
sort key, and a cast whose source already had the target type.

+++ runtime/src/haxe_sys.rs
@@ -4147,7 +4147,7 @@ pub extern "C" fn haxe_objectmap_to_string_typed(
             .map(|(k, v)| {
                 format!(
                     "{} => {}",
-                    map_slot_to_string(*k as u64, key_tag),
+                    map_slot_to_string(*k, key_tag),
                     map_slot_to_string(*v, val_tag)
                 )
             })
